requirements.txt
 bs4
 requests
 lxml
 GitPython